pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ottawa-assistant"
version = "0.1.0"
description = "Ottawa Newcomer Assistant chatbot"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.39.0",
    "langchain>=0.3.0,<1.0",
    "langchain-openai>=0.2.0,<1.0",
    "langchain-ollama>=0.2.0,<1.0",
    "langchain-community>=0.3.0,<1.0",
    "langchain-text-splitters>=0.3.0,<1.0",
    "faiss-cpu>=1.8.0",
    "python-dotenv>=1.0.1",
    "googlesearch-python>=1.2.5",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "pypdf>=4.3.1",
    "tiktoken>=0.7.0",
]

[project.scripts]
ottawa-assistant-ingest = "ottawa_assistant.retriever.ingest:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"

[tool.coverage.run]
source = ["ottawa_assistant"]

[tool.coverage.report]
show_missing = true
